feat: transfer filter logic to a component and configure enums
